chore: windows auto update
